refactor: move `isFunction` to `utils-shared`
