+++ packages/fatcher/src/utils.ts
@@ -1,13 +1,4 @@
 /**
- * Confirm a value whether is a function
- * @param value value to confirm
- * @returns
- */
-export function isFunction(value: unknown): value is (...args: any[]) => any {
-    return typeof value === 'function';
-}
-
-/**
  * Set a plain object to readonly.
  * @param rawData
  * @returns
