+++ packages/utils-shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './getParamsByQuerystring';
 export * from './getQuerystringByParams';
 export * from './getResponseStatus';
+export * from './isFunction';
 export * from './isNodeJS';
 export * from './isRequestWithoutBody';
