+++ packages/utils-shared/src/isFunction.ts
@@ -1,0 +1,8 @@
+/**
+ * Confirm a value whether is a function
+ * @param value value to confirm
+ * @returns
+ */
+export function isFunction(value: unknown): value is (...args: any[]) => any {
+    return typeof value === 'function';
+}
